# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Deltaswap Phylax API\nThis is the API for the Deltaswap Phylax and Explorer.\nThe API has two namespaces: deltaswapscan and phylax.\ndeltaswapscan is the namespace for the explorer and the new endpoints. The prefix is /api/v1.\nphylax is the legacy namespace backguard compatible with phylax node API. The prefix is /v1.\nThis API is public and does not require authentication although some endpoints are rate limited.\nCheck each endpoint documentation for more information.",
        "title": "Deltaswapscan API",
        "termsOfService": "https://wormhole.com/",
        "contact": {
            "name": "API Support",
            "url": "https://discord.com/invite/wormholecrypto",
            "email": "info@wormhole.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/global-tx/{chain_id}/{emitter}/{seq}": {
            "get": {
                "description": "Find a global transaction by VAA ID\nGlobal transactions is a logical association of two transactions that are related to each other by a unique VAA ID.\nThe first transaction is created on the origin chain when the VAA is emitted.\nThe second transaction is created on the destination chain when the VAA is redeemed.\nIf the response only contains an origin tx the VAA was not redeemed.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "find-global-transaction-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the blockchain",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address of the emitter",
                        "name": "emitter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sequence of the VAA",
                        "name": "seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.Tx"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/last-txs": {
            "get": {
                "description": "Returns the number of transactions by a defined time span and sample rate.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "get-last-transactions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time Span, default: 1d, supported values: [1d, 1w, 1mo]. 1mo ​​is 30 days.",
                        "name": "timeSpan",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sample Rate, default: 1h, supported values: [1h, 1d]. Valid configurations with timeSpan: 1d/1h, 1w/1d, 1mo/1d",
                        "name": "sampleRate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/transactions.TransactionCountResult"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/scorecards": {
            "get": {
                "description": "Returns a list of KPIs for Deltaswap.\nTVL is total value locked by token bridge contracts in USD.\nVolume is the all-time total volume transferred through the token bridge in USD.\n24h volume is the volume transferred through the token bridge in the last 24 hours, in USD.\nTotal Tx count is the number of transaction bridging assets since the creation of the network (does not include Pyth or other messages).\n24h tx count is the number of transaction bridging assets in the last 24 hours (does not include Pyth or other messages).\nTotal messages is the number of VAAs emitted since the creation of the network (includes Pyth messages).",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "get-scorecards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.ScorecardsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/token/{chain_id}/{token_address}": {
            "get": {
                "description": "Returns a token symbol, coingecko id and address by chain and token address.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "get-token-by-chain-and-address",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the blockchain",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token address",
                        "name": "token_address",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.Token"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/top-assets-by-volume": {
            "get": {
                "description": "Returns a list of emitter_chain and asset pairs with ordered by volume.\nThe volume is calculated using the notional price of the symbol at the day the VAA was emitted.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "get-top-assets-by-volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time span, supported values: 7d, 15d, 30d.",
                        "name": "timeSpan",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.TopAssetsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/top-chain-pairs-by-num-transfers": {
            "get": {
                "description": "Returns a list of the emitter_chain and destination_chain pair ordered by transfer count.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "get-top-chain-pairs-by-num-transfers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time span, supported values: 7d, 15d, 30d.",
                        "name": "timeSpan",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.TopChainPairsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/transactions/": {
            "get": {
                "description": "Returns transactions. Output is paginated.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "list-transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number. Starts at 0.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of elements per page.",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort results in ascending or descending order.",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter transactions by Address.",
                        "name": "address",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.ListTransactionsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/transactions/{chain_id}/{emitter}/{seq}": {
            "get": {
                "description": "Find VAA metadata by ID.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "get-transaction-by-id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id of the blockchain",
                        "name": "chain_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "address of the emitter",
                        "name": "emitter",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "sequence of the VAA",
                        "name": "seq",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.TransactionDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/api/v1/x-chain-activity": {
            "get": {
                "description": "Returns a list of chain pairs by origin chain and destination chain.\nThe list could be rendered by notional or transaction count.\nThe volume is calculated using the notional price of the symbol at the day the VAA was emitted.",
                "tags": [
                    "deltaswapscan"
                ],
                "operationId": "x-chain-activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time span, supported values: 7d, 30d, 90d, 1y and all-time (default is 7d).",
                        "name": "timeSpan",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Renders the results using notional or tx count (default is notional).",
                        "name": "by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "List of apps separated by comma (default is all apps).",
                        "name": "apps",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/transactions.ChainActivity"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "transactions.AssetWithVolume": {
            "type": "object",
            "properties": {
                "emitterChain": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "symbol": {
                    "type": "string"
                },
                "tokenAddress": {
                    "type": "string"
                },
                "tokenChain": {
                    "$ref": "#/definitions/vaa.ChainID"
                },
                "volume": {
                    "type": "string"
                }
            }
        },
...
        "transactions.ChainActivity": {
            "type": "object",
            "properties": {
                "txs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.Tx"
                    }
                }
            }
        },
...
        "transactions.Destination": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "integer"
                },
                "percentage": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
...
        "transactions.ListTransactionsResponse": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.TransactionDetail"
                    }
                }
            }
        },
...
        "transactions.ScorecardsResponse": {
            "type": "object",
            "properties": {
                "24h_messages": {
                    "description": "Number of VAAs emitted in the last 24 hours (includes Pyth messages).",
                    "type": "string"
                },
                "24h_tx_count": {
                    "description": "Number of VAAs emitted in the last 24 hours (does not include Pyth messages).",
                    "type": "string"
                },
                "24h_volume": {
                    "description": "Volume transferred through the token bridge in the last 24 hours, in USD.",
                    "type": "string"
                },
                "total_tx_count": {
                    "description": "Number of VAAs emitted since the creation of the network (does not include Pyth messages)",
                    "type": "string"
                },
                "total_volume": {
                    "type": "string"
                },
                "tvl": {
                    "description": "Total value locked in USD.",
                    "type": "string"
                }
            }
        },
        "transactions.Token": {
            "type": "object",
            "properties": {
                "coingeckoId": {
                    "type": "string"
                },
                "decimals": {
                    "type": "integer"
                },
                "symbol": {
                    "type": "string"
                }
            }
        },
        "transactions.TopAssetsResponse": {
            "type": "object",
            "properties": {
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.AssetWithVolume"
                    }
                }
            }
        },
        "transactions.TopChainPairsResponse": {
            "type": "object",
            "properties": {
                "chainPairs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.ChainPair"
                    }
                }
            }
        },
        "transactions.TransactionCountResult": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "transactions.Tx": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "integer"
                },
                "destinations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/transactions.Destination"
                    }
                },
                "percentage": {
                    "type": "number"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
...
    }
}